refactor: pulls Praiz's code; updates components
In this code I did the following:
- Pulled Praiz's code which creates a Button component
- Updated the global css to contain general styles for the buttons
- Update classes
- General housekeeping on folder structure
