--- components/btn/Button.jsx
@@ -1,7 +1,0 @@
-export default function Button({ text, submit, className }) {
-  return (
-    <div>
-      <button type={submit ? 'submit' : 'button'} className={className}>{text}</button>
-    </div>
-  );
-}
