+++ components/ui/atoms/btn/Button.jsx
@@ -1,0 +1,9 @@
+export default function Button({ text, submit, className }) {
+  return (
+    <div>
+      <button type={submit ? 'submit' : 'button'} className={className}>
+        {text}
+      </button>
+    </div>
+  );
+}
